+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added Admin Dashboard with `AdminDashboard` using [Blazor](https://dotnet.microsoft.com/en-us/apps/aspnet/web-apps/blazor), [Radzen Table](https://github.com/radzenhq/radzen-blazor) and [MudBlazor](https://github.com/MudBlazor/MudBlazor/) UI. 
+
 ## [4.1.0]
 
 ### Fixed